moved remaining pGet methods to basefullcharacter / character
